gmsol::client

Struct Client

pub struct Client<C> { /* private fields */ }

GMSOL Client.

Implementations

impl<C: Clone + Deref<Target = impl Signer>> Client<C>

pub fn new_with_options( cluster: Cluster, payer: C, options: ClientOptions, ) -> Result<Self>

Create a new Client with the given options.

pub fn new(cluster: Cluster, payer: C) -> Result<Self>

Create a new Client with default options.

pub fn try_clone_with_payer<C2: Clone + Deref<Target = impl Signer>>( &self, payer: C2, ) -> Result<Client<C2>>

Try to clone a new client with a new payer.

pub fn try_clone(&self) -> Result<Self>

Try to clone the client.

pub fn set_subscription_config( &mut self, config: SubscriptionConfig, ) -> &mut Self

Replace the subscription config.

pub fn anchor(&self) -> &Client<C>

Get anchor client.

pub fn program(&self, program_id: Pubkey) -> Program<C>

Create a new Program with the given program id.

pub fn cluster(&self) -> &Cluster

Get the cluster.

pub fn commitment(&self) -> CommitmentConfig

Get the commitment config.

pub fn payer(&self) -> Pubkey

Get the payer.

pub fn rpc(&self) -> &RpcClient

Get RPC Client.

pub fn store_program(&self) -> &Program<C>

Get the store program.

pub fn treasury_program(&self) -> &Program<C>

Get the treasury program.

pub fn timelock_program(&self) -> &Program<C>

Get the timelock program.

pub fn new_store_program(&self) -> Result<Program<C>>

Create a new store program.

pub fn new_treasury_program(&self) -> Result<Program<C>>

Create a new treasury program.

pub fn store_program_id(&self) -> &Pubkey

Get the program id of the store program.

pub fn treasury_program_id(&self) -> &Pubkey

Get the program id of the treasury program.

pub fn timelock_program_id(&self) -> &Pubkey

Get the program id of the timelock program.

pub fn store_transaction(&self) -> TransactionBuilder<'_, C>

Create a transaction builder for the store program.

pub fn treasury_transaction(&self) -> TransactionBuilder<'_, C>

Create a transaction builder for the treasury program.

pub fn timelock_transaction(&self) -> TransactionBuilder<'_, C>

Create a transaction builder for the timelock program.

pub fn bundle_with_options( &self, options: BundleOptions, ) -> BundleBuilder<'_, C>

Create a bundle builder with the given options.

pub fn bundle(&self) -> BundleBuilder<'_, C>

Create a BundleBuilder

pub fn find_store_address(&self, key: &str) -> Pubkey

Find PDA for Store account.

pub fn find_store_wallet_address(&self, store: &Pubkey) -> Pubkey

Find PDA for store wallet account.

pub fn store_event_authority(&self) -> Pubkey

Get the event authority address for the Store program.

pub fn find_market_vault_address( &self, store: &Pubkey, token: &Pubkey, ) -> Pubkey

Find PDA for market vault account.

pub fn find_market_token_address( &self, store: &Pubkey, index_token: &Pubkey, long_token: &Pubkey, short_token: &Pubkey, ) -> Pubkey

Find PDA for market token mint account.

pub fn find_market_address(&self, store: &Pubkey, token: &Pubkey) -> Pubkey

Find PDA for market account.

pub fn find_deposit_address( &self, store: &Pubkey, user: &Pubkey, nonce: &NonceBytes, ) -> Pubkey

Find PDA for deposit account.

pub fn find_first_deposit_owner_address(&self) -> Pubkey

Find first deposit owner address.

pub fn find_withdrawal_address( &self, store: &Pubkey, user: &Pubkey, nonce: &NonceBytes, ) -> Pubkey

Find DPA for withdrawal account.

pub fn find_order_address( &self, store: &Pubkey, user: &Pubkey, nonce: &NonceBytes, ) -> Pubkey

Find PDA for order.

pub fn find_shift_address( &self, store: &Pubkey, owner: &Pubkey, nonce: &NonceBytes, ) -> Pubkey

Find PDA for shift.

pub fn find_position_address( &self, store: &Pubkey, user: &Pubkey, market_token: &Pubkey, collateral_token: &Pubkey, kind: PositionKind, ) -> Result<Pubkey>

Find PDA for position.

pub fn find_claimable_account_address( &self, store: &Pubkey, mint: &Pubkey, user: &Pubkey, time_key: &[u8], ) -> Pubkey

Find claimable account address.

pub fn find_trade_event_buffer_address( &self, store: &Pubkey, authority: &Pubkey, index: u16, ) -> Pubkey

Find trade event buffer address.

pub fn find_user_address(&self, store: &Pubkey, owner: &Pubkey) -> Pubkey

Find User account address.

pub fn find_referral_code_address( &self, store: &Pubkey, code: ReferralCodeBytes, ) -> Pubkey

Find referral code address.

pub fn find_glv_token_address(&self, store: &Pubkey, index: u16) -> Pubkey

Find GLV token address.

pub fn find_glv_address(&self, glv_token: &Pubkey) -> Pubkey

Find GLV address.

pub fn find_glv_deposit_address( &self, store: &Pubkey, owner: &Pubkey, nonce: &NonceBytes, ) -> Pubkey

Find GLV deposit address.

pub fn find_glv_withdrawal_address( &self, store: &Pubkey, owner: &Pubkey, nonce: &NonceBytes, ) -> Pubkey

Find GLV withdrawal address.

pub fn find_gt_exchange_vault_address( &self, store: &Pubkey, time_window_index: i64, time_window: u32, ) -> Pubkey

Find GT exchange vault address.

pub fn find_gt_exchange_address(&self, vault: &Pubkey, owner: &Pubkey) -> Pubkey

Find GT exchange address.

pub fn find_price_feed_address( &self, store: &Pubkey, authority: &Pubkey, index: u16, provider: PriceProviderKind, token: &Pubkey, ) -> Pubkey

Find Custom Price Feed address.

pub fn find_treasury_config_address(&self, store: &Pubkey) -> Pubkey

Find treasury global config address.

pub fn find_treasury_vault_config_address( &self, config: &Pubkey, index: u16, ) -> Pubkey

Find treasury vault config address.

pub fn find_gt_bank_address( &self, treasury_vault_config: &Pubkey, gt_exchange_vault: &Pubkey, ) -> Pubkey

Find GT bank address.

pub fn find_treasury_receiver_address(&self, config: &Pubkey) -> Pubkey

Find treasury receiver address.

pub fn find_timelock_config_address(&self, store: &Pubkey) -> Pubkey

Find timelock config address.

pub fn find_executor_address( &self, store: &Pubkey, role: &str, ) -> Result<Pubkey>

Find executor address.

pub fn find_executor_wallet_address(&self, executor: &Pubkey) -> Pubkey

Find the wallet address of the given executor.

pub async fn get_slot( &self, commitment: Option<CommitmentConfig>, ) -> Result<u64>

Get slot.

pub async fn store_accounts_with_config<T>( &self, filter_by_store: Option<StoreFilter>, other_filters: impl IntoIterator<Item = RpcFilterType>, config: ProgramAccountsConfig, ) -> Result<WithContext<Vec<(Pubkey, T)>>>
where T: AccountDeserialize + Discriminator,

Fetch accounts owned by the Store Program.

pub async fn account_with_config<T>( &self, address: &Pubkey, config: RpcAccountInfoConfig, ) -> Result<WithContext<Option<T>>>
where T: AccountDeserialize,

Fetch account at the given address with config.

The value inside the returned context will be None if the account does not exist.

pub async fn account<T: AccountDeserialize>( &self, address: &Pubkey, ) -> Result<Option<T>>

Fetch account at the given address.

pub async fn store_accounts<T>( &self, filter_by_store: Option<StoreFilter>, other_filters: impl IntoIterator<Item = RpcFilterType>, ) -> Result<Vec<(Pubkey, T)>>
where T: AccountDeserialize + Discriminator,

Fetch accounts owned by the Store Program.

pub async fn store(&self, address: &Pubkey) -> Result<Arc<Store>>

Fetch Store account with its address.

pub async fn user(&self, address: &Pubkey) -> Result<UserHeader>

Fetch user account with its address.

pub async fn authorized_token_map_address( &self, store: &Pubkey, ) -> Result<Option<Pubkey>>

Fetch the TokenMap address of the given store.

pub async fn token_map(&self, address: &Pubkey) -> Result<TokenMap>

Fetch TokenMap account with its address.

pub async fn authorized_token_map(&self, store: &Pubkey) -> Result<TokenMap>

Fetch the authorized token map of the given store.

pub async fn markets_with_config( &self, store: &Pubkey, config: ProgramAccountsConfig, ) -> Result<WithContext<BTreeMap<Pubkey, Market>>>

Fetch all Market accounts of the given store.

pub async fn markets(&self, store: &Pubkey) -> Result<BTreeMap<Pubkey, Market>>

Fetch all Market accounts of the given store.

pub async fn market_with_config<T>( &self, address: &Pubkey, config: RpcAccountInfoConfig, ) -> Result<WithContext<Option<Market>>>

Fetch Market at the given address with config.

The value inside the returned context will be None if the account does not exist.

pub async fn market(&self, address: &Pubkey) -> Result<Arc<Market>>

Fetch Market account with its address.

pub async fn glvs_with_config( &self, store: &Pubkey, config: ProgramAccountsConfig, ) -> Result<WithContext<BTreeMap<Pubkey, Glv>>>

Fetch all Glv accounts of the given store.

pub async fn glvs(&self, store: &Pubkey) -> Result<BTreeMap<Pubkey, Glv>>

Fetch all Glv accounts of the given store.

pub async fn market_status( &self, store: &Pubkey, market_token: &Pubkey, prices: Prices<u128>, maximize_pnl: bool, maximize_pool_value: bool, ) -> Result<MarketStatus>

Fetch MarketStatus with the market token address.

pub async fn market_token_price( &self, store: &Pubkey, market_token: &Pubkey, prices: Prices<u128>, pnl_factor: PnlFactorKind, maximize: bool, ) -> Result<u128>

Fetch current market token price with the market token address.

pub async fn positions( &self, store: &Pubkey, owner: Option<&Pubkey>, market_token: Option<&Pubkey>, ) -> Result<BTreeMap<Pubkey, Position>>

Fetch all Position accounts of the given owner of the given store.

pub async fn position(&self, address: &Pubkey) -> Result<Position>

Fetch Position account with its address.

pub async fn order(&self, address: &Pubkey) -> Result<Order>

Fetch Order account with its address.

pub async fn order_with_config( &self, address: &Pubkey, config: RpcAccountInfoConfig, ) -> Result<WithContext<Option<Order>>>

Fetch Order account at the the given address with config.

The value inside the returned context will be None if the account does not exist.

pub async fn orders( &self, store: &Pubkey, owner: Option<&Pubkey>, market_token: Option<&Pubkey>, ) -> Result<BTreeMap<Pubkey, Order>>

Fetch all Order accounts of the given owner of the given store.

pub async fn deposit(&self, address: &Pubkey) -> Result<Deposit>

Fetch Depsoit account with its address.

pub async fn withdrawal(&self, address: &Pubkey) -> Result<Withdrawal>

Fetch Withdrawal account with its address.

pub async fn price_feed(&self, address: &Pubkey) -> Result<Option<PriceFeed>>

Fetch PriceFeed account with its address.

pub async fn instruction_buffer( &self, address: &Pubkey, ) -> Result<Option<InstructionBuffer>>

Fetch InstructionBuffer account with its address.

pub async fn pub_sub(&self) -> Result<&PubsubClient>

Get the PubsubClient.

pub async fn subscribe_store_cpi_events( &self, commitment: Option<CommitmentConfig>, ) -> Result<impl Stream<Item = Result<WithSlot<Vec<StoreCPIEvent>>>>>

Subscribe to StoreCPIEvents from the store program.

pub async fn historical_store_cpi_events( &self, address: &Pubkey, commitment: Option<CommitmentConfig>, ) -> Result<impl Stream<Item = Result<WithSlot<Vec<StoreCPIEvent>>>>>

Fetch historical StoreCPIEvents for the given account.

pub async fn complete_order( &self, address: &Pubkey, commitment: Option<CommitmentConfig>, ) -> Result<Option<TradeEvent>>

Wait for an order to be completed using current slot as min context slot.

pub async fn last_order_events( &self, order: &Pubkey, before_slot: u64, commitment: CommitmentConfig, ) -> Result<Vec<StoreCPIEvent>>

Get last order events.

pub async fn complete_order_with_config( &self, address: &Pubkey, slot: u64, polling: Duration, commitment: Option<CommitmentConfig>, ) -> Result<Option<TradeEvent>>

Wait for an order to be completed with the given config.

pub async fn shutdown(&self) -> Result<()>

Shutdown the client gracefully.

pub async fn gt_exchanges( &self, store: &Pubkey, owner: &Pubkey, ) -> Result<BTreeMap<Pubkey, GtExchange>>

Get GT exchanges.

impl<C: Deref<Target = impl Signer> + Clone> Client<C>

pub fn create_first_glv_deposit( &self, store: &Pubkey, glv_token: &Pubkey, market_token: &Pubkey, ) -> CreateGlvDepositBuilder<'_, C>

Create first GLV deposit.

impl<C: Deref<Target = impl Signer> + Clone> Client<C>

pub fn create_first_deposit( &self, store: &Pubkey, market_token: &Pubkey, ) -> CreateDepositBuilder<'_, C>

Create first deposit.