gmsol

Module constants

GMSOL constants.

Modules

glv

Default GLV configs.

market

Default market configs.

Constants

CLAIMABLE_ACCOUNT_SEED

Claimable Account Seed.

DEFAULT_ADL_PRICES_MAX_STALENESS

Default max ADL prices staleness (in seconds).

DEFAULT_BORROWING_FEE_ABOVE_OPTIMAL_USAGE_FACTOR_FOR_LONG

Default borrowing fee above optimal usage factor for long.

DEFAULT_BORROWING_FEE_ABOVE_OPTIMAL_USAGE_FACTOR_FOR_SHORT

Default borrowing fee above optimal usage factor for short.

DEFAULT_BORROWING_FEE_BASE_FACTOR_FOR_LONG

Default borrowing fee base factor for long.

DEFAULT_BORROWING_FEE_BASE_FACTOR_FOR_SHORT

Default borrowing fee base factor for short.

DEFAULT_BORROWING_FEE_EXPONENT_FOR_LONG

Default borrowing fee exponent for long.

DEFAULT_BORROWING_FEE_EXPONENT_FOR_SHORT

Default borrowing fee exponent for short.

DEFAULT_BORROWING_FEE_FACTOR_FOR_LONG

Default borrowing fee factor for long.

DEFAULT_BORROWING_FEE_FACTOR_FOR_SHORT

Default borrowing fee factor for short.

DEFAULT_BORROWING_FEE_OPTIMAL_USAGE_FACTOR_FOR_LONG

Default borrowing fee optimal usage factor for long.

DEFAULT_BORROWING_FEE_OPTIMAL_USAGE_FACTOR_FOR_SHORT

Default borrowing fee optimal usage factor for short.

DEFAULT_CLAIMABLE_TIME_WINDOW

Default claimable time window.

DEFAULT_FUNDING_FEE_DECREASE_FACTOR_PER_SECOND

Default funding fee decrease factor per second.

DEFAULT_FUNDING_FEE_EXPONENT

Default funding fee exponent.

DEFAULT_FUNDING_FEE_FACTOR

Default funding factor.

DEFAULT_FUNDING_FEE_INCREASE_FACTOR_PER_SECOND

Default funding fee increase factor per second.

DEFAULT_FUNDING_FEE_MAX_FACTOR_PER_SECOND

Default funding fee max factor per second.

DEFAULT_FUNDING_FEE_MIN_FACTOR_PER_SECOND

Default funding fee min factor per second.

DEFAULT_FUNDING_FEE_THRESHOLD_FOR_DECREASE_FUNDING

Default funding fee threshold for decrease funding.

DEFAULT_FUNDING_FEE_THRESHOLD_FOR_STABLE_FUNDING

Default funding fee threshold for stable funding.

DEFAULT_GLV_MAX_SHIFT_PRICE_IMPACT_FACTOR

Default GLV max shift price impact factor.

DEFAULT_GLV_MIN_SHIFT_INTERVAL_SECS

Default GLV min shift interval seconds.

DEFAULT_GLV_MIN_SHIFT_VALUE

Default GLV min shift value.

DEFAULT_GT_VAULT_TIME_WINDOW

Default GT vault time window size.

DEFAULT_IGNORE_OPEN_INTEREST_FOR_USAGE_FACTOR

Default ignore open interest for usage factor.

DEFAULT_LIQUIDATION_FEE_FACTOR

Default liquidation fee factor.

DEFAULT_MAX_NEGATIVE_POSITION_IMPACT_FACTOR

Default max negative position impact factor.

DEFAULT_MAX_OPEN_INTEREST_FOR_LONG

Default max open interest for long.

DEFAULT_MAX_OPEN_INTEREST_FOR_SHORT

Default max open interest for short.

DEFAULT_MAX_PNL_FACTOR_FOR_LONG_ADL

Default max pnl factor for long adl.

DEFAULT_MAX_PNL_FACTOR_FOR_LONG_DEPOSIT

Default max pnl factor for long deposit.

DEFAULT_MAX_PNL_FACTOR_FOR_LONG_TRADER

Default max pnl factor for long trader.

DEFAULT_MAX_PNL_FACTOR_FOR_LONG_WITHDRAWAL

Default max pnl factor for long withdrawal.

DEFAULT_MAX_PNL_FACTOR_FOR_SHORT_ADL

Default max pnl factor for short adl.

DEFAULT_MAX_PNL_FACTOR_FOR_SHORT_DEPOSIT

Default max pnl factor for short deposit.

DEFAULT_MAX_PNL_FACTOR_FOR_SHORT_TRADER

Default max pnl factor for short trader.

DEFAULT_MAX_PNL_FACTOR_FOR_SHORT_WITHDRAWAL

Default max pnl factor for short withdrawal.

DEFAULT_MAX_POOL_AMOUNT_FOR_LONG_TOKEN

Default max pool amount for long token.

DEFAULT_MAX_POOL_AMOUNT_FOR_SHORT_TOKEN

Default max pool amount for short token.

DEFAULT_MAX_POOL_VALUE_FOR_DEPOSIT_LONG_TOKEN

Default max pool value for deposit for long token.

DEFAULT_MAX_POOL_VALUE_FOR_DEPOSIT_SHORT_TOKEN

Default max pool value for deposit for short token.

DEFAULT_MAX_POSITION_IMPACT_FACTOR_FOR_LIQUIDATIONS

Default max position impact factor for liquidations.

DEFAULT_MAX_POSITIVE_POSITION_IMPACT_FACTOR

Default max positive position impact factor.

DEFAULT_MIN_COLLATERAL_FACTOR

Default min collateral factor.

DEFAULT_MIN_COLLATERAL_FACTOR_FOR_OPEN_INTEREST_FOR_LONG

Default min collateral factor for open interest for long

DEFAULT_MIN_COLLATERAL_FACTOR_FOR_OPEN_INTEREST_FOR_SHORT

Default min collateral factor for open interest for short

DEFAULT_MIN_COLLATERAL_VALUE

Default min collateral value.

DEFAULT_MIN_PNL_FACTOR_AFTER_LONG_ADL

Default min pnl factor after long adl.

DEFAULT_MIN_PNL_FACTOR_AFTER_SHORT_ADL

Default min pnl factor after short adl.

DEFAULT_MIN_POSITION_IMPACT_POOL_AMOUNT

Default min position impact pool amount.

DEFAULT_MIN_POSITION_SIZE_USD

Default min position size usd.

DEFAULT_MIN_TOKENS_FOR_FIRST_DEPOSIT

Default min tokens for first deposit.

DEFAULT_OPEN_INTEREST_RESERVE_FACTOR

Default open interest reserve factor.

DEFAULT_ORACLE_MAX_AGE

Default oracle max age.

DEFAULT_ORACLE_MAX_FUTURE_TIMESTAMP_EXCESS

Default oracle max future timestamp excess (in seconds).

DEFAULT_ORACLE_MAX_TIMESTAMP_RANGE

Default oracle max timestamp range.

DEFAULT_ORACLE_REF_PRICE_DEVIATION

Default oracle ref price deviation.

DEFAULT_ORDER_FEE_FACTOR_FOR_NEGATIVE_IMPACT

Default order fee factor for negative impact.

DEFAULT_ORDER_FEE_FACTOR_FOR_POSITIVE_IMPACT

Default order fee factor for positive impact.

DEFAULT_POSITION_IMPACT_DISTRIBUTE_FACTOR

Default position impact distribute factor.

DEFAULT_POSITION_IMPACT_EXPONENT

Default position impact exponent.

DEFAULT_POSITION_IMPACT_NEGATIVE_FACTOR

Default position impact negative factor.

DEFAULT_POSITION_IMPACT_POSITIVE_FACTOR

Default position impact positive factor.

DEFAULT_RECEIVER_FACTOR

Default receiver factor.

DEFAULT_RECENT_TIME_WINDOW

Default recent time window.

DEFAULT_REQUEST_EXPIRATION

Default request expiration.

DEFAULT_RESERVE_FACTOR

Default reserve factor.

DEFAULT_SKIP_BORROWING_FEE_FOR_SMALLER_SIDE

Default skip borrowing fee for smaller side.

DEFAULT_SWAP_FEE_FACTOR_FOR_NEGATIVE_IMPACT

Default swap fee factor for negative impact.

DEFAULT_SWAP_FEE_FACTOR_FOR_POSITIVE_IMPACT

Default swap fee factor for positive impact.

DEFAULT_SWAP_IMPACT_EXPONENT

Default swap impact exponent.

DEFAULT_SWAP_IMPACT_NEGATIVE_FACTOR

Default swap impact negative factor.

DEFAULT_SWAP_IMPACT_POSITIVE_FACTOR

Default swap impact positive factor.

ESCROW_ACCOUNT_SEED

Escrow Account Seed.

EVENT_AUTHORITY_SEED

Event authority SEED.

FUNDING_AMOUNT_PER_SIZE_ADJUSTMENT

Adjustment factor for saving funding amount per size.

GT_MINT_SEED

GT Mint Seed.

MAREKT_TOKEN_MINT_SEED

Market Token Mint Address Seed.

MARKET_DECIMALS

Decimals of usd values of factors.

MARKET_TOKEN_DECIMALS

Decimals of a market token.

MARKET_USD_TO_AMOUNT_DIVISOR

USD value to amount divisor.

MARKET_USD_UNIT

Unit USD value i.e. one.

MARKET_VAULT_SEED

Market Vault Seed.