gmsol

Module types

GMSOL types.

Modules

common

Common types.

deposit

Deposit.

feature

Feature.

glv

GLV.

gt

GT.

market

This module defines the Market struct and its associated state.

oracle

Oracle.

order

Order.

position

Position.

roles

Roles.

shift

Shift.

store

Store.

store_instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

timelock

Re-export gmsol_timelock types.

token_config

Token Config.

treasury

Re-export gmsol_treasury types.

user

User.

withdrawal

Withdrawal.

Structs

Addresses

Addresses.

Amounts

Amounts.

BorrowingFeesUpdated

Market borrowing fees updated event.

Chainlink

The Chainlink Program.

Deposit

Deposit.

DepositCreated

Deposit Created Event.

DepositExecuted

Deposit executed Event.

DepositRemoved

Deposit removed event.

EventClocks

Market clocks.

EventOtherState

Market State.

EventPool

A pool for market.

EventPositionState

Position State.

EventTradeFees

Trade Fees.

EventTradeOutputAmounts

Output amounts.

EventTradePnl

Trade PnL.

EventTradePrice

Price.

EventTradePrices

Trade Prices.

EventTransferOut

Transfer Out.

Factors

Factors.

FeedConfig

Price Feed Config.

Glv

Glv.

GlvDeposit

Glv Deposit.

GlvDepositRemoved

GLV Deposit removed event.

GlvPricing

GLV pricing event.

GlvShift

Glv Shift.

GlvWithdrawal

Glv Withdrawal.

GlvWithdrawalRemoved

GLV Withdrawal removed event.

GtUpdated

GT updated event.

Market

Market.

MarketFeesUpdated

Market fees updated event.

MarketMeta

Market Metadata.

MarketStateUpdated

Market State Updated Event.

Members

Fixed size map generated by the macro.

Oracle

Oracle Account.

OracleFlagContainer

Flags container generated by the macro.

Order

Order.

OrderActionParams

Order params.

OrderCreated

Order created event.

OrderRemoved

Order removed event.

OtherState

Market State.

PoolStorage

A pool storage for market.

Position

Position.

PositionDecreased

Position decrease event.

PositionIncreased

Position increased event.

PriceFeed

Custom Price Feed.

PriceFeedPrice

Price structure for Price Feed.

PriceValidator

Price Validator.

Pyth

The Pyth receiver program.

RoleKey

The key of a Role.

RoleMap

Fixed size map generated by the macro.

RoleMetadata

Role Metadata.

RoleStore

Roles Store.

Shift

Shift.

ShiftActionParams

ShiftRemoved

Shift removed event.

ShiftTokenAccounts

Store

Data Store.

SwapExecuted

Swap executed Event.

Switchboard

The Switchboard receiver program.

TokenConfig

TokenMap

Token Map.

TokenMapHeader

Header of TokenMap.

TokenMapMut

Mutable Reference to Token Map.

TokenMapRef

Reference to Token Map.

Tokens

Fixed size map generated by the macro.

TradeData

Trade event data.

TradeEvent

Trade event.

TradeFees

Trade Fees.

TradeFlagContainer

Flags container generated by the macro.

TradeOutputAmounts

Output amounts.

TradePnl

Trade PnL.

TradePrice

Price.

TradePrices

Prices.

Treasury

Treasury.

UpdateOrderParams

Update Order Params.

UpdateTokenConfigParams

UserHeader

Header of User Account.

Withdrawal

Withdrawal.

WithdrawalCreated

Withdrawal created event.

WithdrawalExecuted

Withdrawal executed Event.

WithdrawalRemoved

Withdrawal removed event.

Enums

AddressKey

Address keys.

AmountKey

Amount keys.

FactorKey

Factor keys.

GlvPricingKind

Pricing kind.

GtUpdateKind

GT Update Kind.

MarketConfigKey

Market config keys.

PriceProviderKind

Supported Price Provider Kind.

TradeFlag

Trade Data Flags.

Constants

DEFAULT_HEARTBEAT_DURATION

Default heartbeat duration for price updates.

DEFAULT_PRECISION

Default precision for price.

DEFAULT_TIMESTAMP_ADJUSTMENT

Default timestamp adjustment.

MAX_MEMBERS

Max number of members.

MAX_ROLES

Max number of roles.

MAX_ROLE_NAME_LEN

Max length of the role anme.

Traits

Event

Event.

HasMarketMeta

Type that has market meta.

Seed

Data type that has SEED.

TokenMapAccess

Read Token Map.

TokenMapAccessMut

Token Map Operations.

TokenMapLoader

Token Map Loader.

ValidateOracleTime

Validate Oracle Time.

ValidateOracleTimeExt

Extension trait for ValidateOracleTime.

Type Aliases

Amount

Factor

NonceBytes

Nonce Bytes.