gmsol::types::store

Struct Store

#[repr(C)]
pub struct Store {
    pub authority: Pubkey,
    pub token_map: Pubkey,
    /* private fields */
}

Data Store.

Fields
authority: Pubkey

Store authority.

token_map: Pubkey

The token map to used.

Implementations

impl Store

pub const MAX_LEN: usize = 32usize

Maximum length of key.

pub const WALLET_SEED: &'static [u8] = b"store_wallet"

Wallet Seed.

pub fn init( &mut self, authority: Pubkey, key: &str, bump: u8, receiver: Pubkey, holding: Pubkey, ) -> Result<(), Error>

Initialize.

pub fn role(&self) -> &RoleStore

Get the role store.

pub fn key(&self) -> Result<&str, Error>

Get the key of the store.

pub fn enable_role(&mut self, role: &str) -> Result<(), Error>

Enable a role.

pub fn disable_role(&mut self, role: &str) -> Result<(), Error>

Disable a role.

pub fn has_role(&self, authority: &Pubkey, role: &str) -> Result<bool, Error>

Check if the roles has the given enabled role. Returns true only when the role is enabled and the roles has that role.

Note

If the cluster has restarted, this function returns true if and only if the authority has the RESTART_ADMIN role.

pub fn grant(&mut self, authority: &Pubkey, role: &str) -> Result<(), Error>

Grant a role.

pub fn revoke(&mut self, authority: &Pubkey, role: &str) -> Result<(), Error>

Revoke a role.

pub fn is_authority(&self, authority: &Pubkey) -> bool

Check if the given pubkey is the authority of the store.

pub fn has_admin_role(&self, authority: &Pubkey) -> Result<bool, Error>

Check if the given authority has the ADMIN role.

Note

If the cluster has restarted, addresses with the RESTART_ADMIN role also have the ADMIN role.

pub fn token_map(&self) -> Option<&Pubkey>

Get token map address.

pub fn get_amount(&self, key: &str) -> Result<&u64, Error>

Get amount.

pub fn get_amount_by_key(&self, key: AmountKey) -> Option<&u64>

Get amount by key.

pub fn get_amount_mut(&mut self, key: &str) -> Result<&mut u64, Error>

Get amount mutably

pub fn get_factor(&self, key: &str) -> Result<&u128, Error>

Get factor.

pub fn get_factor_by_key(&self, key: FactorKey) -> Option<&u128>

Get factor by key.

pub fn get_factor_mut(&mut self, key: &str) -> Result<&mut u128, Error>

Get factor mutably

pub fn get_address(&self, key: &str) -> Result<&Pubkey, Error>

Get address.

pub fn get_address_by_key(&self, key: AddressKey) -> Option<&Pubkey>

Get address by key.

pub fn get_address_mut(&mut self, key: &str) -> Result<&mut Pubkey, Error>

Get address mutably

pub fn request_expiration_at(&self, start: i64) -> Result<i64, CoreError>

Calculate the request expiration time.

pub fn claimable_time_window(&self) -> Result<NonZero<u64>, Error>

Get claimable time window size.

pub fn claimable_time_window_index(&self, timestamp: i64) -> Result<i64, Error>

Get claimable time window index for the given timestamp.

pub fn claimable_time_key(&self, timestamp: i64) -> Result<[u8; 8], Error>

Get claimable time key for the given timestamp.

pub fn holding(&self) -> &Pubkey

Get holding address.

pub fn validate_claim_fees_address(&self, address: &Pubkey) -> Result<(), Error>

Validate whether fees can be claimed by this address.

pub fn receiver(&self) -> Pubkey

Get the receiver address.

pub fn next_receiver(&self) -> Pubkey

Get the next receiver address.

pub fn gt(&self) -> &GtState

Get GT State.

pub fn get_feature_disabled( &self, domain: DomainDisabledFlag, action: ActionDisabledFlag, ) -> Option<bool>

Get feature disabled.

pub fn is_feature_disabled( &self, domain: DomainDisabledFlag, action: ActionDisabledFlag, ) -> bool

Is the given feature disabled.

pub fn validate_feature_enabled( &self, domain: DomainDisabledFlag, action: ActionDisabledFlag, ) -> Result<(), Error>

Validate whether the given features is enabled.

pub fn has_restarted(&self) -> Result<bool, Error>

Returns whether the cluster has restarted since last update.

pub fn validate_not_restarted(&self) -> Result<&Store, Error>

Validate the cluster has not restarted.

pub fn validate_not_restarted_mut(&mut self) -> Result<&mut Store, Error>

Validate the cluster has not restarted for mutable reference.

pub fn order_fee_discount_factor( &self, rank: u8, is_referred: bool, ) -> Result<u128, Error>

Get order fee discount factor.