gmsol::types

Module store_instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

AcceptReceiver

Instruction.

AcceptReferralCode

Instruction.

AcceptStoreAuthority

Instruction.

AutoDeleverage

Instruction.

CancelOrderIfNoPosition

Instruction.

CancelReferralCodeTransfer

Instruction.

CheckAdmin

Instruction.

CheckRole

Instruction.

ClaimFeesFromMarket

Instruction.

ClearAllPrices

Instruction.

CloseDeposit

Instruction.

CloseEmptyClaimableAccount

Instruction.

CloseGlvDeposit

Instruction.

CloseGlvShift

Instruction.

CloseGlvWithdrawal

Instruction.

CloseGtExchange

Instruction.

CloseMarketConfigBuffer

Instruction.

CloseOrder

Instruction.

CloseShift

Instruction.

CloseWithdrawal

Instruction.

ConfirmGtExchangeVault

Instruction.

CreateDeposit

Instruction.

CreateGlvDeposit

Instruction.

CreateGlvShift

Instruction.

CreateGlvWithdrawal

Instruction.

CreateOrder

Instruction.

CreateShift

Instruction.

CreateWithdrawal

Instruction.

DisableRole

Instruction.

EnableRole

Instruction.

ExecuteDecreaseOrder

Instruction.

ExecuteDeposit

Instruction.

ExecuteGlvDeposit

Instruction.

ExecuteGlvShift

Instruction.

ExecuteGlvWithdrawal

Instruction.

ExecuteIncreaseOrSwapOrder

Instruction.

ExecuteShift

Instruction.

ExecuteWithdrawal

Instruction.

GetMarketStatus

Instruction.

GetMarketTokenPrice

Instruction.

GrantRole

Instruction.

GtSetExchangeTimeWindow

Instruction.

GtSetOrderFeeDiscountFactors

Instruction.

GtSetReferralRewardFactors

Instruction.

HasAdmin

Instruction.

HasRole

Instruction.

Initialize

Instruction.

InitializeGlv

Instruction.

InitializeGt

Instruction.

InitializeMarket

Instruction.

InitializeMarketConfigBuffer

Instruction.

InitializeMarketVault

Instruction.

InitializeOracle

Instruction.

InitializePriceFeed

Instruction.

InitializeReferralCode

Instruction.

InitializeTokenMap

Instruction.

InsertAddress

Instruction.

InsertAmount

Instruction.

InsertFactor

Instruction.

InsertGlvMarket

Instruction.

InsertOrderFeeDiscountForReferredUser

Instruction.

IsTokenConfigEnabled

Instruction.

Liquidate

Instruction.

MarketTransferIn

Instruction.

MigrateReferralCode

Instruction.

PrepareAssociatedTokenAccount

Instruction.

PrepareGtExchangeVault

Instruction.

PreparePosition

Instruction.

PrepareTradeEventBuffer

Instruction.

PrepareUser

Instruction.

PushToMarketConfigBuffer

Instruction.

PushToTokenMap

Instruction.

PushToTokenMapSynthetic

Instruction.

RemoveGlvMarket

Instruction.

RequestGtExchange

Instruction.

RevokeRole

Instruction.

SetExpectedProvider

Instruction.

SetFeedConfig

Instruction.

SetMarketConfigBufferAuthority

Instruction.

SetPricesFromPriceFeed

Instruction.

SetReferrer

Instruction.

SetTokenMap

Instruction.

ToggleFeature

Instruction.

ToggleGlvMarketFlag

Instruction.

ToggleGtMinting

Instruction.

ToggleMarket

Instruction.

ToggleTokenConfig

Instruction.

TokenDecimals

Instruction.

TokenExpectedProvider

Instruction.

TokenFeed

Instruction.

TokenName

Instruction.

TokenPrecision

Instruction.

TokenTimestampAdjustment

Instruction.

TransferReceiver

Instruction.

TransferReferralCode

Instruction.

TransferStoreAuthority

Instruction.

UpdateAdlState

Instruction.

UpdateGlvConfig

Instruction.

UpdateGlvMarketConfig

Instruction.

UpdateLastRestartedSlot

Instruction.

UpdateMarketConfig

Instruction.

UpdateMarketConfigFlag

Instruction.

UpdateMarketConfigWithBuffer

Instruction.

UpdateOrder

Instruction.

UpdatePriceFeedWithChainlink

Instruction.

UseClaimableAccount

Instruction.