gmsol::types

Module timelock

Re-export gmsol_timelock types.

Modules

config

Timelock Config.

executor

Executor.

instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

utils

Utils for using instruction buffer.

Structs

Executor

Executor.

ExecutorWalletSigner

Executor Wallet Signer.

InstructionAccount

Instruction Account.

InstructionFlagContainer

Flags container generated by the macro.

InstructionHeader

Instruction Header.

InstructionRef

Reference to the instruction.

TimelockConfig

Timelock Config.

Enums

InstructionAccountFlag

Flags of Instruction Accounts.

InstructionFlag

Flags of Instruction.

Traits

InstructionAccess

Instruction Access.

InstructionLoader

Instruction Loader.

Functions

create_executor_wallet_pda

Create executor wallet PDA.

find_executor_wallet_pda

Find executor wallet PDA.