gmsol::types::treasury

Module instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

CancelSwap

Instruction.

ClaimFees

Instruction.

CompleteGtExchange

Instruction.

ConfirmGtBuyback

Instruction.

CreateSwap

Instruction.

DepositToTreasuryVault

Instruction.

InitializeConfig

Instruction.

InitializeTreasuryVaultConfig

Instruction.

InsertTokenToTreasuryVault

Instruction.

PrepareGtBank

Instruction.

RemoveTokenFromTreasuryVault

Instruction.

SetBuybackFactor

Instruction.

SetGtFactor

Instruction.

SetReferralReward

Instruction.

SetTreasuryVaultConfig

Instruction.

SyncGtBank

Instruction.

ToggleTokenFlag

Instruction.

TransferReceiver

Instruction.

WithdrawFromTreasuryVault

Instruction.