gmsol::utils::builder::oracle

Trait PullOraclePriceConsumer

pub trait PullOraclePriceConsumer {
    // Required methods
    fn feed_ids(&mut self) -> impl Future<Output = Result<FeedIds>>;
    fn process_feeds(
        &mut self,
        provider: PriceProviderKind,
        map: FeedAddressMap,
    ) -> Result<()>;
}

Pull Oracle Price Consumer.

Required Methods

fn feed_ids(&mut self) -> impl Future<Output = Result<FeedIds>>

Returns a reference to tokens and their associated feed IDs that require price updates.

fn process_feeds( &mut self, provider: PriceProviderKind, map: FeedAddressMap, ) -> Result<()>

Processes the feed address map returned from the pull oracle.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.