gmsol::utils::rpc

Module accounts

Account RPC utils.

Structs

ProgramAccountsConfig

Program Accounts Config.

ProgramAccountsConfigForRpc

Program Accounts Config.

Functions

account_with_context

Return the decoded account at the given address, along with context.

accounts_lazy_with_context

Returns all program accounts of the given type matching the specified filters as an iterator, along with context. Deserialization is executed lazily.

get_account_with_context

Get account with context.

get_program_accounts_with_context

Get program accounts with context.

get_token_accounts_by_owner_with_context

Get token accounts by owner and return with the context.